requirements.txt
transformers
pandas
numpy
faiss-cpu
sentence_transformers
gdown
pickle